pyproject.toml
[project]
name = "lang-chain"
version = "0.1.0"
description = "LangChain Complete Guide"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.0",
    "fastembed>=0.7.3",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.0",
    "python-dotenv>=1.2.1",
]
